3863 PipelineVisitor
