+++ src/pageEditor/testHelpers.ts
@@ -27,6 +27,7 @@ import runtimeSlice from "@/pageEditor/slices/runtimeSlice";
 import { logSlice } from "@/components/logViewer/logSlice";
 import { createRenderWithWrappers } from "@/testUtils/testHelpers";
 import validationListenerMiddleware from "@/pageEditor/validation/validationListenerMiddleware";
+import analysisSlice from "@/analysis/analysisSlice";
 
 const renderWithWrappers = createRenderWithWrappers(() =>
   configureStore({
@@ -40,6 +41,7 @@ const renderWithWrappers = createRenderWithWrappers(() =>
       savingExtension: savingExtensionSlice.reducer,
       runtime: runtimeSlice.reducer,
       logs: logSlice.reducer,
+      analysis: analysisSlice.reducer,
       // This api reducer may be needed at some point, but it's not mocked properly yet, so
       //  we're not including it for now, until it becomes an issue.
       // [appApi.reducerPath]: appApi.reducer,
